simple-rabbitmq-example/requirements.txt
paho-mqtt ~= 2.1
